moving all flicker code to flicker sytem
